Move ResponseStructures to Model package
